Delete Authorization

+++ App/controllers/ListingController.php
@@ -4,9 +4,12 @@
 namespace App\Controllers;
 
 use Framework\Database;
+use Framework\Session;
 use PDO;
 use Framework\Validation;
 use ValueError;
+use Framework\Authorization;
+
 
 class ListingController
 {
@@ -23,7 +26,7 @@ class ListingController
      */
     public function index()
     {
-        $listings = $this->db->query("SELECT * FROM listings LIMIT 6")->fetchAll(PDO::FETCH_OBJ);
+        $listings = $this->db->query("SELECT * FROM listings ORDER BY created_at DESC LIMIT 6")->fetchAll(PDO::FETCH_OBJ);
         loadView('listings/index', ['listings' => $listings]);
     }
 
@@ -90,7 +93,7 @@ class ListingController
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -162,9 +165,16 @@ class ListingController
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        // Check if listing exists
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if (!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You are not authorized to delete this listing';
+            return redirect('/listings/' . $listing->id);
         }
 
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
